+++ build.sh
@@ -3,6 +3,10 @@
 #
 # This script is built with commands available to Git Bash on Windows. (mingw32)
 #
+# This script and scripts in resources/ should be identical in the 
+# https://github.com/keymanapp/keyboards and 
+# https://github.com/keymanapp/keyboards_starter repos
+#
 
 function display_usage {
   echo "Usage: $0 [-validate] [-codesign] [-start] [-s] [-d] [-c] [-w] [-t project_target] [target]"
@@ -42,7 +46,18 @@ KEYBOARDINFO_SCHEMA_DIST_JSON="$KEYBOARDROOT/tools/keyboard_info.distribution.js
 parse_args "$@"
 
 #
+# Are we in the keyboards_starter repo?
 #
+
+if [[ -d "$KEYBOARDROOT/template" ]]; then
+  if [[ -d "$KEYBOARDROOT/release" ]]; then
+    die "This repo should not have both a /release/ folder and a /template/ folder. The /template/ folder should be present only in the keyboards_starter repo."
+  fi
+  KEYBOARDS_STARTER=1
+fi
+
+#
+# Select action
 #
 
 if [ "$DO_BUILD" = false ]; then
@@ -57,7 +72,12 @@ fi
 # Collect filenames
 #
 
-KEYBOARD_INFO_PATHS="$KEYBOARDROOT"/*/*/*/*.keyboard_info
+if [[ $KEYBOARDS_STARTER == 1 ]]; then
+  KEYBOARD_INFO_PATHS="$KEYBOARDROOT"/*/*/*/*/*.keyboard_info
+else
+  KEYBOARD_INFO_PATHS="$KEYBOARDROOT"/*/*/*/*.keyboard_info
+fi
+
 KEYBOARD_INFOS=($KEYBOARD_INFO_PATHS)
 KEYBOARD_INFOS=`printf -- '%s\n' "${KEYBOARD_INFOS[@]}"`
 
@@ -65,20 +85,28 @@ KEYBOARD_INFOS=`printf -- '%s\n' "${KEYBOARD_INFOS[@]}"`
 # Run build
 #
 
-if [[ "$TARGET" ]]; then
-  if [[ "$TARGET" == */* ]] && [[ (-d "$TARGET") ]]; then
-    group=$(cut -d / -f 1 <<< "$TARGET")
-    echo "--- Only building $group $TARGET ---"
-    build_keyboard $group "$TARGET"
-  elif [[ "$TARGET" == "release" ]] || [[ "$TARGET" == "legacy" ]] || [[ "$TARGET" == "experimental" ]]; then
-    # Assuming release|legacy|experimental
-    echo "--- Only building $TARGET ---"
-    build_keyboards "$TARGET"
+if [[ $KEYBOARDS_STARTER == 1 ]]; then
+  if [[ "$TARGET" ]]; then
+    build_keyboard "$TARGET"
   else
-    display_usage
+    build_keyboards
   fi
 else
-  build_keyboards release
-  build_keyboards legacy
-  build_keyboards experimental
+  if [[ "$TARGET" ]]; then
+    if [[ "$TARGET" == */* ]] && [[ (-d "$TARGET") ]]; then
+      group=$(cut -d / -f 1 <<< "$TARGET")
+      echo "--- Only building $group $TARGET ---"
+      build_keyboard $group "$TARGET"
+    elif [[ "$TARGET" == "release" ]] || [[ "$TARGET" == "legacy" ]] || [[ "$TARGET" == "experimental" ]]; then
+      # Assuming release|legacy|experimental
+      echo "--- Only building $TARGET ---"
+      build_keyboards "$TARGET"
+    else
+      display_usage
+    fi
+  else
+    build_keyboards release
+    build_keyboards legacy
+    build_keyboards experimental
+  fi
 fi