+++ legacy/g/grkpoly2/source/help/grkpoly2.php
@@ -241,7 +241,7 @@ for full details on how to type all the letters, diacritics, and punctuation of 
 <h2>Troubleshooting</h2>
 <p class='keymanweb'>If the characters you type or those on the On Screen Keyboard do not appear to display correctly, please read the <a target="_blank" href="/troubleshooting/#boxes">KeymanWeb troubleshooting guide</a>.</p>
 
-<p>For any other questions, <a target="_blank" href="http://www.tavultesoft.com/contact.php">contact us</a>.</p>
+<p>For any other questions, <a target="_blank" href="https://keyman.com/contact/">contact us</a>.</p>
 
 </div>
 <div id="Resources">
@@ -260,10 +260,10 @@ for full details on how to type all the letters, diacritics, and punctuation of 
 </div>
 <div>
 <h3>Other Internet Resources</h3>
-<p><a target="_blank" href="http://keymankeyboards.com/?id=321">Download the Greek Polytonic Unicode Keyboard for Keyman Desktop</a></p>
+<p><a target="_blank" href="https://keyman.com/keyboards/grkpoly2">Download the Greek Polytonic Unicode Keyboard for Keyman Desktop</a></p>
 <p><a target="_blank" href="http://www.omniglot.com/writing/greek.htm">More on the Greek Writing System at Omniglot.com</a><br/>
-<a target="_blank" href="http://en.wikipedia.org/wiki/Greek_alphabet">More on the Greek Writing System at Wikipedia.org</a></p>
-<p><a target="_blank" href="http://en.wikipedia.org/wiki/Greek_language">More on the Greek Language at Wikipedia.org</a></p>
+<a target="_blank" href="https://en.wikipedia.org/wiki/Greek_alphabet">More on the Greek Writing System at Wikipedia.org</a></p>
+<p><a target="_blank" href="https://en.wikipedia.org/wiki/Greek_language">More on the Greek Language at Wikipedia.org</a></p>
 </div>
 </div>
 <div id="Technical">