+++ legacy/h/hararab/source/help/hararab.php
@@ -1372,7 +1372,8 @@ vowels</h3>
 <p>&copy; 2009 State Library of Victoria and the Australian Saay Harari Association </p>
 <div id="contact">
 <h2>contact</h2>
-<p id="name"><a href="http://home.vicnet.net.au/~andrewc/" target="_blank">Andrew Cunningham</a><br />
+<p><a href="mailto:enabling.languages@gmail.com">Andrew Cunningham</a></p>
+<p>
 Vicnet, State Library of Victoria<br />
 328 Swanston St.<br />
 Melbourne VIC 3000<br />