+++ legacy/kbd/kbdcz/source/help/kbdcz.php
@@ -1,0 +1,26 @@
+<?php
+  $pagename = 'Czech Basic Keyboard Help';
+  $pagetitle = 'Czech Basic Keyboard Help';
+  require_once('header.php');
+?>
+<style type='text/css'>
+    h1 { font: bold 16pt Tahoma; color: #4444cc }
+    h2 { font: bold 12pt Tahoma; color: #4444cc }
+    span.key { font: bold 10pt Verdana; padding: 0 4px; border: solid 1px #aaaaaa; margin: 0 4px; }
+</style>
+
+<p>This keyboard layout is designed for Czech.  It includes an on screen keyboard
+which can be viewed by clicking on the Keyman icon and selecting the On Screen Keyboard
+menu item.</p>
+
+	  <img width="100%" src="keyboard_layout.png">
+
+
+<h2>Release History</h2>
+<p>Version 1.0, released 10 November 2008</p>
+
+<h2>Links</h2>
+<ul>
+<li>Home Page: <a href='https://keyman.com/keyboards/kbdcz'>https://keyman.com/keyboards/kbdcz</a></li>
+<li><a href='https://keyman.com/keyboards/languages/cs'>Other keyboards for these languages</a></li>
+</ul>