+++ legacy/r/russian-tajik/source/help/russian-tajik.php
@@ -1,0 +1,370 @@
+﻿<!DOCTYPE html>
+<?php
+  $pagename = 'Cyrillic Russian+Tajik Keyboard Help';
+  $pagetitle = 'Cyrillic Russian+Tajik Keyboard Help';
+  require_once('header.php');
+?>
+
+<body>
+  <p>Keyboard &#169; 2003 Linguasoft</p>
+   <section id='overview'>
+	<h2>Overview</h2>
+    <p>This keyboard is designed to type Tajik Cyrillic and also includes most combinations required to type Russian.</p>
+	<h4>Fonts</h4>
+	<p>This is a Unicode keyboard and works with any Unicode font which has support for Cyrillic script. </p>
+   </section>
+   
+<h2>The Tajik Cyrillic &amp; Arabic-Persian alphabet</h2>
+<h2 style='margin-left: 200px'>-алифбои тоҷикӣ - الفبای تاجکی</h2>
+ 
+<table class='alphabet'> 
+<tr>
+  <td><b>Keyman Input for Cyrillic</b></td>
+  <td><b>Cyr_upper</b></td>
+  <td><b>Cyr_lower</b></td>
+  <td><b>Cyr_xlit</b></td>
+  <td><b>Arabic-Persian</b></td>
+  <td><b>Ar_xlit</b></td>
+</tr>
+<tr >
+  <td  rowspan=2 >A, a</td>
+  <td  rowspan=2 >А</td>
+  <td  rowspan=2 >а</td>
+  <td  rowspan=2 >a</td>
+  <td>َ</td>
+  <td>a</td>
+</tr>
+<tr >
+  <td>ه</td>
+  <td>&#234;*(h)</td>
+</tr>
+<tr >
+  <td>B, b</td>
+  <td>Б</td>
+  <td>б</td>
+  <td>b</td>
+  <td>ب</td>
+  <td>b</td>
+</tr>
+<tr >
+  <td>V, v</td>
+  <td>В</td>
+  <td>в</td>
+  <td>v</td>
+  <td>و</td>
+  <td>v(ū/&#251;)</td>
+</tr>
+<tr >
+  <td>G, g</td>
+  <td>Г</td>
+  <td>г</td>
+  <td>g</td>
+  <td>گ</td>
+  <td>g</td>
+</tr>
+<tr >
+  <td>\G, \g</td>
+  <td>Ғ</td>
+  <td>ғ</td>
+  <td>ġ</td>
+  <td>غ</td>
+  <td>ġ</td>
+</tr>
+<tr >
+  <td>D, d</td>
+  <td>Д</td>
+  <td>д</td>
+  <td>d</td>
+  <td>د</td>
+  <td>d</td>
+</tr>
+<tr >
+  <td>E, e <i>or</i> Je, je <i>or</i> `E, `e</td>
+  <td>Е</td>
+  <td>е</td>
+  <td>e</td>
+  <td>ی</td>
+  <td>e(y/ī)</td>
+</tr>
+<tr >
+  <td>Zh, zh <i>or</i> ^Z, ^z</td>
+  <td>Ж</td>
+  <td>ж</td>
+  <td>ž</td>
+  <td>ژ</td>
+  <td>ž</td>
+</tr>
+<tr >
+  <td  rowspan=4 >Z, z</td>
+  <td  rowspan=4 >З</td>
+  <td  rowspan=4 >з</td>
+  <td  rowspan=4 >z</td>
+  <td>ز</td>
+  <td>z</td>
+</tr>
+<tr >
+  <td>ض</td>
+  <td>ż</td>
+</tr>
+<tr >
+  <td>ظ</td>
+  <td>ẓ</td>
+</tr>
+<tr >
+  <td>ذ</td>
+  <td>ẕ</td>
+</tr>
+<tr >
+  <td>I, i</td>
+  <td>И</td>
+  <td>и</td>
+  <td>i</td>
+  <td>ِ</td>
+  <td>i</td>
+</tr>
+<tr >
+  <td>\I, \i</td>
+  <td>Ӣ</td>
+  <td>ӣ</td>
+  <td>ī</td>
+  <td>ی</td>
+  <td>ī(y/e)</td>
+</tr>
+<tr >
+  <td>J, j</td>
+  <td>Й</td>
+  <td>й</td>
+  <td>j</td>
+  <td>ی</td>
+  <td>y(ī/e)</td>
+</tr>
+<tr >
+  <td>K, k</td>
+  <td>К</td>
+  <td>к</td>
+  <td>k</td>
+  <td>ک</td>
+  <td>k</td>
+</tr>
+<tr >
+  <td>\K, \k</td>
+  <td>Қ</td>
+  <td>қ</td>
+  <td>q</td>
+  <td>ق</td>
+  <td>q</td>
+</tr>
+<tr >
+  <td>L, l</td>
+  <td>Л</td>
+  <td>л</td>
+  <td>l</td>
+  <td>ل</td>
+  <td>l</td>
+</tr>
+<tr >
+  <td>M, m</td>
+  <td>М</td>
+  <td>м</td>
+  <td>m</td>
+  <td>م</td>
+  <td>m</td>
+</tr>
+<tr >
+  <td>N, n</td>
+  <td>Н</td>
+  <td>н</td>
+  <td>n</td>
+  <td>ن</td>
+  <td>n</td>
+</tr>
+<tr >
+  <td>O, o</td>
+  <td>О</td>
+  <td>о</td>
+  <td>o</td>
+  <td>ا</td>
+  <td>&#244;(')</td>
+</tr>
+<tr >
+  <td>P, p</td>
+  <td>П</td>
+  <td>п</td>
+  <td>p</td>
+  <td>پ</td>
+  <td>p</td>
+</tr>
+<tr >
+  <td>R, r</td>
+  <td>Р</td>
+  <td>р</td>
+  <td>r</td>
+  <td>ر</td>
+  <td>r</td>
+</tr>
+<tr >
+  <td  rowspan=3>S, s</td>
+  <td  rowspan=3>С</td>
+  <td  rowspan=3>с</td>
+  <td  rowspan=3>s</td>
+  <td>س</td>
+  <td>s</td>
+</tr>
+<tr >
+  <td>ص</td>
+  <td>ṣ</td>
+</tr>
+<tr >
+  <td>ث</td>
+  <td>s̱</td>
+</tr>
+
+<tr >
+  <td  rowspan=2>T, t</td>
+  <td  rowspan=2>Т</td>
+  <td  rowspan=2>т</td>
+  <td  rowspan=2>t</td>
+  <td>ت</td>
+  <td>t</td>
+</tr>
+<tr >
+  <td>ط</td>
+  <td>ṭ</td>
+</tr>
+<tr >
+  <td>U, u</td>
+  <td>У</td>
+  <td>у</td>
+  <td>u</td>
+  <td>ُ</td>
+  <td>u</td>
+</tr>
+<tr >
+  <td>\U, \u</td>
+  <td>Ӯ</td>
+  <td>ӯ</td>
+  <td>ū</td>
+  <td>و</td>
+  <td>ū(v/&#251;)</td>
+</tr>
+<tr >
+  <td>F, f</td>
+  <td>Ф</td>
+  <td>ф</td>
+  <td>f</td>
+  <td>ف</td>
+  <td>f</td>
+</tr>
+<tr >
+  <td>X, x</td>
+  <td>Х</td>
+  <td>х</td>
+  <td>ẖ</td>
+  <td>خ</td>
+  <td>ḫ</td>
+</tr>
+<tr >
+  <td  rowspan=2>\X, \x</td>
+  <td  rowspan=2>Ҳ</td>
+  <td  rowspan=2>ҳ</td>
+  <td  rowspan=2>h</td>
+  <td>ح</td>
+  <td>ḥ</td>
+</tr>
+<tr >
+  <td>ه</td>
+  <td>h(&#234;)</td>
+</tr>
+<tr>
+  <td>Ch, ch <i>or</i> ^C, ^c <i>or</i> H, h</td>
+  <td>Ч</td>
+  <td>ч</td>
+  <td>č</td>
+  <td>چ</td>
+  <td>č</td>
+</tr>
+<tr >
+  <td>\H, \h</td>
+  <td>Ҷ</td>
+  <td>ҷ</td>
+  <td>ǧ</td>
+  <td>ج</td>
+  <td>j</td>
+</tr>
+<tr >
+  <td>Sh, sh <i>or</i> W, w</td>
+  <td>Ш</td>
+  <td>ш</td>
+  <td>&#353;</td>
+  <td>ش</td>
+  <td>&#353;</td>
+</tr>
+<tr >
+  <td>~ (<i>after another letter</i>)</td>
+  <td>Ъ</td>
+  <td>ъ</td>
+  <td>ʿ</td>
+  <td>ع</td>
+  <td>ʿ</td>
+</tr>
+<tr >
+  <td>^E, ^e</td>
+  <td>Э</td>
+  <td>э</td>
+  <td>&#232;</td>
+  <td>(?)</td>
+  <td>(?)</td>
+</tr>
+<tr >
+  <td>Ja, ja <i>or</i> `A, `a</td>
+  <td>Я</td>
+  <td>я</td>
+  <td>ja</td>
+  <td>(يَ)</td>
+  <td>(ya)</td>
+</tr>
+<tr >
+  <td>Ju, ju <i>or</i> `U, `u</td>
+  <td>Ю</td>
+  <td>ю</td>
+  <td>ju</td>
+  <td>(يُ)</td>
+  <td>(yu)</td>
+</tr>
+<tr >
+  <td>Jo, jo <i>or</i> `O, `o</td>
+  <td>Ё</td>
+  <td>ё</td>
+  <td>jo</td>
+  <td>(يا)</td>
+  <td>(y&#244;)</td>
+</tr>
+<tr >
+  <td>Shch, shch <i>or</i> Q, q</td>
+  <td>Щ</td>
+  <td>щ</td>
+  <td colspan=3 rowspan=3 ><i>(These Cyrillic letters aren&#8217;t used in Tajik)</i></td>
+</tr>
+<tr >
+  <td>Y, y</td>
+  <td>Ы</td>
+  <td>ы</td>
+</tr>
+<tr >
+  <td>` (<i>after another letter</i>)</td>
+  <td>Ь</td>
+  <td>ь</td>
+</tr>
+</table>
+<p style='margin-left: 400px'>*) &#234; = -a [&#230;] <i>(at the end of a word)</i></p>
+
+
+<h2>Version History</h2>
+
+
+
+<p>1.0 - Initial release</p>
+<p>1.1 - Added support for "ьо" sequence with by typing "```o"; packaged for Keyman 8+ with on screen keyboard and language association.</p>
+<p>1.2 - Added support for "жь" sequence with by typing "zh`" (also caps); packaged for Keyman 8+ with on screen keyboard and language association.</p>
+
+</body>