+++ release/e/ekwtamil99uni/source/help/ekwtamil99uni.php
@@ -25,7 +25,7 @@
   ?>
 
   
-<p style='margin:0px'>Keyboard &#169; 2008 thamizha.com and Tavultesoft</p>
+<p style='margin:0px'>Keyboard &#169; 2008, 2015, 2018 thamizha.com and Tavultesoft</p>
 
 <br/>
 <a href="#Overview">Overview</a><br/>
@@ -326,7 +326,7 @@ This keyboard was created by Mugunth (mugunth@gmail.com), Umar (csd_one@yahoo.co
 </p>
 <h3>Copyright and Terms of Use</h3>
 <p>
-The Tamil99 Keyboard for Keyman Desktop and KeymanWeb is Copyright 2008 thamizha.com and Tavultesoft.  It may be freely distributed and used, but must not be modified or adapted in any way without written permission from Tavultesoft.
+The Tamil99 Keyboard for Keyman Desktop and KeymanWeb is Copyright 2008, 2015, 2018 thamizha.com and Tavultesoft.  It may be freely distributed and used, but must not be modified or adapted in any way without written permission from Tavultesoft.
 </p>
 <p>
 THIS SOFTWARE IS PROVIDED BY TAVULTESOFT PTY LTD "AS IS" AND ANY