+++ release/n/nw_iranian_latin/source/help/nw_iranian_latin.php
@@ -723,7 +723,11 @@ This is a Latin keyboard for transcription of NW Iranian languages.
 </table>
 
       <H2>License</H2>
+<<<<<<< HEAD
       <P>NW Iranian Latin Keyman Keyboard is copyright ©2017-19 SIL 
+=======
+      <P>NW Iranian Latin Keyman Keyboard is copyright ©2017-2018 SIL 
+>>>>>>> master
       International. It is distributed under The MIT License (MIT)</P>
       <DIV class=Note>
 <p>Permission is hereby granted, free of charge, to any person obtaining a copy