+++ release/packages/galaxie_greek_hebrew_positional/source/help/galaxie_greek_hebrew_positional.php
@@ -1,0 +1,21 @@
+﻿<?php
+  $pagename = 'Galaxie Greek/Hebrew (Positional) Keyboard Package Help';
+  $pagetitle = 'Galaxie Greek/Hebrew (Positional) Keyboard Package Help';
+  require_once('header.php');
+?>
+
+<p>This package includes 2 positional keyboards called: Galaxie Greek (Phonetic) and 
+Galaxie Hebrew (Positional).</p>
+
+<p>The Hebrew keyboard is designed for Biblical Hebrew, a Semitic language of the ancient 
+Near East and the language of the Hebrew Bible.</p>
+
+<p>The Greek keyboard is designed for Koine Greek, a trade language of the ancient Mediterranean 
+and the language of the New Testament and Septuagint Scriptures.</p>
+
+  <section id='related'>
+	<h2>Keyboard Layouts</h2>
+<ul><a href="https://help.keyman.com/keyboard/galaxie_hebrew_positional">Galaxie Hebrew (Positional)</a> </ul>
+<ul><a href="https://help.keyman.com/keyboard/galaxie_greek_positional">Galaxie Greek (Phonetic)</a> </ul>
+  </section>
+