+++ release/rac/rac_palula/source/help/rac_palula.php
@@ -1,0 +1,99 @@
+<?php 
+  $pagename = 'Rachitrali-Palula Keyman Keyboard Help';
+  $pagetitle = $pagename;
+  require_once('header.php');
+?>
+<h1>The Rachitrali-Palula Keyman Keyboard Package</h1>
+      <P><SPAN class=author_date_hits>Rehmat Aziz Chitrali, 
+      2018-06-03</SPAN></P>
+
+<h2>Introduction</h2>
+<p>This keyboard is designed for the Palula language spoken in Chitral Pakistan in its standard Perso-Arabic script, as used Pakistan. It is based on phonetic keyboard for Palula Language developed by Pakistani linguist, author, researcher, scientist,  Inventor and Software Developer Mr.Rehmat Aziz Chitrali for Summer Institute of Linguistics(SIL), and later approved by Khowar Academy(an literary Association working on the language and culture documentation in Northern Pakistan) used in Windows 8.1 and above as Palula Keyboard.
+</p>
+
+<p>This is the first Palula language keyboard created by Rehmat Aziz Chitrali, Pakistani linguist, Scientist,  Inventor,  Software Developer and Researcher. This keyboard is for Palula Language based on the keyboard layout as developed by Mr.Rehmat Aziz Chitrali and approved by Khowar Academy of Pakistan as standard keyboard for Palula language.</p>
+
+<p>The Palula Keyboard Project(PKP) was supervised and developed by Rehmat Aziz Chitrali, who heads the Khowar Academy Chitral Pakistan.</p>
+
+<p>Khowar Academy Chitral was established in April 1996 specifically to promote the endangered  languages of Chitral Pakistan. The Academy has worked extensively on Palula linguistics, standardization for computing, and modeling of script, speech and language. Research is being conducted in all aspects of Palula, including acoustic phonetics, phonology, morphology, syntax, grammar and semantics.</p>
+
+<p/>
+<h2>Font</h2>
+<p style="text-align: justify;">
+The Rachitrali-Palula Keyman keyboard is designed to work with any extended 
+Arabic script font.  
+</p>
+
+
+<h2>Keyboard</h2>
+<p>
+This keyboard layout is designed by Rehmat Aziz Chitrali for the Palula language. It includes 
+an on screen keyboard which can be viewed by clicking on the Keyman icon 
+and selecting the On Screen Keyboard menu item.     </p>
+
+
+      <H2>Keyboard Layout</H2>
+      <P>The keyboard layout is displayed below in Keyboard graphics.</P>
+      	
+      	<h3>Base Version</h3>
+      	<img src="rac_palula base.png" alt="Base Version" style="width:600px;">
+
+      	<h3>Shift Version</h3>
+      	<img src="rac_palula shift.png" alt="Shift Version" style="width:600px;">
+      	
+<h2>How to use the Rachitrali-Palula Keyboard? Please read the following:-</h2>
+
+<p style="text-align: right; dir=rtl; font-size:200%">ہمارے متعلق
+</p>
+<p style="text-align: right; dir=rtl; font-size:200%">پالولہ زبان
+</p>
+<p style="text-align: right; dir=rtl; font-size:150%">پالولہ زبان پاکستان کے صوبہ خیبر پختونخوا کی وادی عشریت، بیوڑی، پوریگل اور شیشی کوہ میں بولی جاتی ہے اس زبان میں جہاں رحمت عزیز چترالی کا پالولہ قاعدہ، پالولہ حروف تہجی، پالولہ خوشخطی، پالولہ سافٹ ویر، پالولہ اردو ذخیرہء الفاظ اور پالولہ ٹیکسٹ ایڈیٹر منظر عام پہ آ چکے ہیں وہاں پالولہ زبان بولنے والے نوجوانوں نے بھی اپنی مادری زبان پالولہ میں کتابیں لکھنے کا عزم ظاہر کیا ہے۔ پالولہ زبان کی اپنی مخصوص اصوات بھی ہیں جوکہ کھوار، پالولہ اور پشتو زبان کی اصوات سے مماثلت رکھتی ہیں اور یہ وہ مخصوص آوازیں ہیں جوکہ اردو اور دیگر پاکستانی زبانوں میں موجود نہیں۔ پالولہ زبان کی ان مخصوص آوازوں کے لیے سب سے پہلے پاکستان کے ضلع چترال سے تعلق رکھنے والے ماہر لسانیات اور محقق جناب رحمت عزیز چترالی ۱۹۹۶ نے چند حروف ایجاد کیے اور کھوار اکیڈمی کے قیام کے بعد انہوں نے پالولہ زبان کے ساتھ ساتھ دیگر پاکستانی زبانوں کو کمپیوٹر پر تحریر کے قابل بنانے کے لیے کیبورڈ بنانے کا سلسلہ شروع کیا اور پاکستان کی چالیس زبانوں کو ایک ہی کیبورڈ کے زریعے تحریر کے قابل بناکر عالمی ریکارڈ قائم کرکے پاکستان کا نام عالمی سطح پر روشن کیا وادی چترال میں بولی جانے والی زبان پالولہ کے لیے کیبورڈ بھی ایجاد کرنے کا سہرہ آپ ہی کے سر جاتا ہے اور آپ کے ایجاد کردہ کلیدی تختی کی مدد سے پالولہ زبان میں تحریر کا سلسلہ جاری ہے اور رحمت عزیز چترالی کا یہ کام پالولہ زبان کی تاریخ میں ایک سنگ میل کی حیثیت رکھتا ہے، پاکستانی زبانوں کے لیے بے لوث خدمات انجام دینے پر  ہم حکومت پاکستان سے رحمت عزیز چترالی کے لیے ''ایوارڈ برائے حسن کارکردگی'' کی سفارش کرتے ہیں امید ہے کہ حکومت پاکستان رحمت عزیز چترالی اور ان جیسے دیگر  پاکستانی ماہرین لسانیات اور ادیبوں  کے لیے ایوارڈو اعزازات کا اعلان کرے گی، رحمت عزیز چترالی نے پہلی بار کھوار زبان اور پالولہ زبان کو گوگل کیبورڈ اور جی بورڈ میں شامل کرکے ان معدومیت کا شکار زبانوں کو بچانے کے لیے اہم کردار ادا کیا جو کہ زبانوں کی تاریخ میں سنہرے حروف سے لکھا جائے گا
+</p>
+
+
+<p style="text-align: right; dir=rtl; font-size:200%">
+پالولہ کیبورڈ کے متعلق
+</p>
+<p style="text-align: right; dir=rtl; font-size:150%">
+پالولہ کیبورڈ پالولہ زبان کا پہلا سافٹ ویر اور کیبورڈ ہے جسے کھوار اکیڈمی چترال سے منسلک ماہر لسانیات رحمت عزیز چترالی نے ایجاد کیا ہے  اس کلیدی تختی میں ڈاکومینٹیشن کے زریعے کیبورڈ کے استعمال کے طریقہء کار کی وضاحت کی گئی ہے، اس کیبورڈ کے تین ورژن ہیں پہلا ورژن'' بیس ورژن'' کہلاتا ہے دوسرا ورژن ''شفٹ ورژن'' ہے اور تیسرا ورژن ''کنٹرول آلٹر ورژن'' کہلاتا ہے، رحمت عزیز چترالی نے یہ پالولہ کیبورڈ اور پالولہ سافٹ ویر کھوار اکیڈمی کے تعاون و اشتراک سے سمر انسٹیٹیوٹ آف لنگوسٹکس کے لیے تیار کیا ہے جوکہ آپ کی پاکستان میں بولی جانے والی معدومیت کا شکار زبان پالولہ کے لیے ایک بہت بڑے کارنامے اور احسان سے کم نہیں، رحمت عزیز چترالی کا یہ کارنامہ ہے کہ انہوں نے پالولہ زبان کے اس سافٹ ویر اور کیبورڈ کو مایکروسافٹ کی ممبر کمپنی ٹاول سافٹ، سمر انسٹیٹوٹ آف لنگوسٹکس اور کیمین کے ساتھ رجسٹر اور ریلیز کروایا اور  اس پالولہ سافٹ ویر کو استفادہء عام کے لیے کھوار اکیڈمی نے ٹاول سافٹ اور کیمین کے تعاون و اشتراک سے ریلیز کردیا ہے امید ہے کہ یہ سافٹ ویر آپ کو پسند آئے گا</p>
+</p>
+<p style="text-align: right; dir=rtl; font-size:200%">
+ہمارے متعلق
+</p>
+<p style="text-align: right; dir=rtl; font-size:150%">
+پالولہ سافٹ ویر اور پالولہ ٹیکسٹ ایڈیٹر ایجاد کرنے کا تاریخی کام پاکستان کے ضلع چترال سے تعلق رکھنے والے لسانی ماہر جناب رحمت عزیز چترانی نے شروع کیا ہے۔ اور آج اس سافٹ ویر کے زریعے رحمت عزیز چترالی نے پالولہ جیسی غیر تحریری زبان کو تحریری زبان میں تبدیل کرکے کمپیوٹر اور موبائل کی زبان بنا دیا ہے اور پالولہ بولنے والے لوگ اب فیس بک اور ٹویٹر میں بھی اس سافٹ ویر کے زریعے پالولہ لکھ رہے ہیں جو کہ پالولہ دان طبقے کے لیے یقینا بہت بڑی خوشی کی بات ہے
+</p>
+
+      <P class=Caption></P><A name=license></A>
+      <H2>License</H2>
+      <P>Rachitrali-Palula Keyman Keyboard is copyright © 2018 Rehmat Aziz Chitrali. 
+      	It is distributed under The MIT License (MIT)</P>
+      <DIV class=Note>
+<p>Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:</p>
+
+<p>The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.</p>
+
+<p>THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.</p></DIV>
+
+
+
+<h2>Release History</h2>
+<p>Version 1.0, released 06 June 2018</p>
+
+<h2>Links</h2>
+<ul>
+<li><a href='http://keymankeyboards.com/iso:phl'>Other keyboards for these languages</a></li>
+</ul>
+