+++ release/s/sabdalipi_assamese/source/help/sabdalipi_assamese.php
@@ -1,0 +1,86 @@
+<?php
+  $pagename = 'Assamese (Sabdalipi) Keyboard Help';
+  $pagetitle = 'Assamese (Sabdalipi) Keyboard Help';
+  require_once('header.php');
+?>
+
+<div id='Overview'>
+		<!-- A brief introduction to the keyboard, including intended users and font/hardware requirements.  This is basically
+the info available on the keyboard download site.  Instructions on using the keyboard DO NOT go here. -->
+  <h2>Overview</h2>
+  <p>
+  This keyboard is designed for the Assamese language.  It is a phonetic keyboard, which means the characters are arranged according to the letters of the English keyboard.
+  </p>
+</div>
+
+
+
+<div id='Using'>
+  <h2>Using this Keyboard</h2>
+  <h3>Keyboard Layout</h3>
+  		<!-- On Screen Keyboard is included here -->
+  <h4>Default state</h4>
+  <p><img src="sabdalipi_assameseU_.png" alt="Default keyboard state"/></p>
+  <h4>Shift state</h4>
+  <p><img src="sabdalipi_assameseU_S.png" alt="Shift keyboard state"/></p>
+</div>
+
+<div id="Examples">
+<h3>Examples</h3>
+		<!-- Simple example(s) illustrating the notes in the Quickstart section above -->
+
+	<table cellspacing=4 style='border-style:solid; border-width:1px; border-color:#8080FF; background-color:#E0E0FF; padding:4px'>
+		<col width=100px />
+		<col width=150px />
+<p style='margin:0px'>&#169; 2009-2018 TI Trade</p>
+		<col width=150px />
+		<col width=350px />
+
+		<tr style='color:#0000AF; font-weight:bold'>
+			<td>Language</td>
+			<td>Phrase</td>
+			<td>Meaning</td>
+			<td>Type the following keys</td>
+		</tr>
+		<tr style='color:#0000AF; font-weight:normal'>
+			<td>Assamese</td>
+			<td><span class='lang2' style='color:#0000FF; font-size:150%'>অসমীয়া</span></td>
+			<td>"Assamese"</td>
+			<td><kbd>A</kbd><kbd>s</kbd><kbd>m</kbd><kbd>E</kbd><kbd>x</kbd><kbd>a</kbd></td>
+		</tr>
+	</table>
+
+
+</div>
+
+<div id="Documentation">
+<h3>Full Documentation</h3>
+		<!-- The guts of the documentation: reasoning behind the keyboard layout; detailed instructions on typing & editing; instructions on special characters or keyboard behaviour, etc.; complete character/input chart or link to chart; -->
+
+
+<p>The Sabdalipi Assamese keyboard uses a phonetic-style layout, so that the key sequence
+for each Assamese (Unicode) character is usually the similar sounding English character, for example:</p>
+
+<p style='margin-left:30px'><kbd>k</kbd> is <span class='lang2'>ক</span>, <kbd>g</kbd> is <span class='lang2'>গ</span>, <kbd>p</kbd> is <span class='lang2'>প</span> and so on.</p>
+<p>Similarly uppercase characters are used for the deep sounding characters, for example:</p>
+<p style='margin-left:30px'> <kbd>K</kbd> is <span class='lang2'>খ</span>, <kbd>G</kbd> is <span class='lang2'>ঘ</span>, <kbd>P</kbd> is <span class='lang2'>ফ</span> and so on.</p>
+
+<p>The plus '<kbd>+</kbd>' sign is used to create juktyakhars, for example:</p>
+
+<p style='margin-left:30px'>typing '<kbd>l</kbd><kbd>+</kbd><kbd>p</kbd>' would result in <span class='lang2'>ল্প</span><br />
+typing '<kbd>j</kbd><kbd>+</kbd><kbd>j</kbd><kbd>+</kbd><kbd>b</kbd>' would result in <span class='lang2'>জ্জ্ব</span></p>
+
+<p>A special key is reserved for all kaar jogs, for example:</p>
+<p style='margin-left:30px'><kbd>a</kbd> stands for Aakaar<br />
+<kbd>u</kbd> stands for Ukaar, i.e <kbd>p</kbd><kbd>u</kbd> gives <span class='lang2'>পু</span>, <kbd>r</kbd><kbd>u</kbd> gives <span class='lang2'>ৰু</span> and <kbd>g</kbd><kbd>u</kbd> gives <span class='lang2'>গু</span><br />
+<kbd>y</kbd> stands for rhokaar, i.e <kbd>g</kbd><kbd>y</kbd> is <span class='lang2'>গ্ৰ</span>, <kbd>t</kbd><kbd>y</kbd> is <span class='lang2'>ত্ৰ</span> and <kbd>k</kbd><kbd>y</kbd> is <span class='lang2'>ক্ৰ</span></p>
+
+<p>The keyboard layout and kaar jogs are shown below.</p>
+
+<img src='kbdlayout.jpg' /> <img src='Kaarjogs.jpg' />
+
+</div>
+
+<hr>
+
+<p style='margin:0px'>&#169; 2009-2018 TI Trade</p>