+++ release/sil/sil_lisu_basic/source/help/sil_lisu_basic.php
@@ -1,0 +1,12 @@
+<?php
+  $pagename = 'Lisu Basic (SIL) Keyboard Help';
+  $pagetitle = 'Lisu Basic (SIL) Keyboard Help';
+  require_once('header.php');
+?>
+
+  <p>This keyboard is a mnemonic Lisu keyboard layout following a QWERTY key arrangement.</p>
+  <h2>Keyboard layout</h2>
+  <h3>Default (Unshifted) state</h3>
+  <p><img src="sil_lisu_basicU_.png" alt="Lisu Basic Keyboard: Unshifted (default) state"></p>
+    <h3>Shift state</h3>
+    <p><img src="sil_lisu_basicU_S.png" alt="Lisu Basic Keyboard: Shifted state"></p>