+++ release/j/jorai/HISTORY.md
@@ -1,5 +1,13 @@
 Jorai Keyboard Change History
 =======================
+1.0.2 (27 June 2018)
+--------------------
+* Include the correct touch keyboard in the keyboard package
+
+1.0.1 (21 May 2018)
+-----------------
+* Changed language tag to include Latn
+* Adjusted spacing on OSK to better see diacritics
 
 1.0 (10 March 2018)
 -----------------