+++ release/n/nw_iranian_latin/HISTORY.md
@@ -1,9 +1,12 @@
 NW Iranian Latin Change History
 ===============================
-
-1.0.1 (25 Feb 2019)
+1.0.2 (25 Feb 2019)
 -------------------
-Removed Balochi language codes
+Removed language codes
+
+1.0.1 (18 Jun 2018)
+-------------------
+Add touch keyboard to package
 
 1.0 (13 Jun 2017)
 -----------------