+++ release/o/osage_nation/HISTORY.md
@@ -1,6 +1,16 @@
 Osage Keyboard Change History
 =======================
 
+1.2.1 (19 Oct 2018)
+-----------------
+* Changed to use .kvk (compiled OSK) in the .kps file
+
+1.2 (14 May 2018)
+-----------------
+* Changed touch layout key for U+104E6
+* Changed to use .kvks format
+* Adjust font information and BCP-47 tag
+
 1.1 (23 Apr 2018)
 -----------------
 * Added font for support on mobile devices