+++ release/t/tainua/HISTORY.md
@@ -1,7 +1,13 @@
 Tai Nua Change History
 =======================
+1.0.2 (18 Jun 2018)
+-------------------
+* Add touch keyboard to package
+
+1.0.1 (4 Jun 2018)
+------------------
+* Added script information to language metadata in .kps
 
 1.0 (20 Oct 2016)
 -----------------
-
 * Created by Sai Ai Wong